% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report.R
\name{analysis.report}
\alias{analysis.report}
\title{Create an analysis report}
\usage{
analysis.report(
  interface = T,
  data,
  year = 2019,
  author = "Cliff Clavin",
  email = "cliff@fws.gov",
  phone = "555-867-5309",
  output_file = "report.pdf",
  output_dir = getwd(),
  ...
)
}
\arguments{
\item{interface}{if \code{TRUE}, user inputs the report parameters using a shiny interface}

\item{data}{input data}

\item{year}{a numeric value indicating the year of the desired report}

\item{author}{the name of the report author}

\item{output_file}{output file name in \link{render}. Default is "report.html".}

\item{output_dir}{output directory for report in \link{render}. Default is user's current directory.}

\item{\dots}{other arguments to be passed to \link{render}.}
}
\value{

}
\description{
Create an analysis report
}
