% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{annual.estimates}
\alias{annual.estimates}
\title{Create Plots of Annual Estimates of Species Occurrence and Detection Probability and LaTeX Code for Analysis Report}
\usage{
annual.estimates(Epsi, Ep, year, figdir.path)
}
\arguments{
\item{Epsi}{Data.frame containing estimated values for occurrence probabilities, SE,
upper & lower CLs, and corresponding habitat covariate values.}

\item{Ep}{Data.frame containing estimated values for detection probabilities, SE,
upper & lower CLs, and corresponding visibility covariate values.}

\item{year}{Numeric value for survey year.}

\item{figdir.path}{Directory path for temporary storage of report figures.}
}
\value{
List of length equal to number of survey years processed. Each element is contains a NULL value.
}
\description{
Estimates are based on top-ranked model and are conditional on estimated
relationships between model parameters and covariates when applicable.
}
