% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_functions.R
\name{annual.summaries}
\alias{annual.summaries}
\title{Create Year-Specific Summary Table and Figures and LaTeX Code for QC Report}
\usage{
annual.summaries(occu, site, year, unmkd, naiveoccu, habcov, figdir.path)
}
\arguments{
\item{occu}{Data.frame containing raw species occurrence data.}

\item{site}{Data.frame containing raw site-level data}

\item{year}{Numeric value for survey year.}

\item{unmkd}{Object of class unmarkedFrameOccu used by unmarked package to fit single-season occupancy model.}

\item{naiveoccu}{Vector containing values of 0 or 1 for each surveyed site indicating whether
the species was observed (1) or unobserved (0).}

\item{habcov}{A RasterLayer object containing habitat covariate values.}

\item{figdir.path}{Directory path for temporary storage of report figures.}
}
\value{
List of length equal to number of survey years processed. Each element is contains a NULL value.
}
\description{
Create Year-Specific Summary Table and Figures and LaTeX Code for QC Report
}
