% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{annual.tables}
\alias{annual.tables}
\title{Create Model Selection and Parameter Estimate Tables and LaTeX Code for Analysis Report}
\usage{
annual.tables(modsel.table, year)
}
\arguments{
\item{modsel.table}{Data.frame containing model output from \code{unmarked}}

\item{year}{Numeric value for survey year.}
}
\value{
List of length equal to number of survey years processed. Each element is contains a NULL value.
}
\description{
Create Model Selection and Parameter Estimate Tables and LaTeX Code for Analysis Report
}
