% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{betas.plot}
\alias{betas.plot}
\title{Plot Estimates and CIs of Linear Model Coefficients for Species Occurrence and Detection Probalities}
\usage{
betas.plot(betas, years, modranks)
}
\arguments{
\item{betas}{List of length equal to number of survey years.  Each element contains a ist of length 2
containing estimates of model parameters for (1) occurrence probability and (2) detection probability}

\item{years}{numeric vector containing survey years}

\item{modranks}{dataframe containing model rankings for all survey years}
}
\description{
Plot Estimates and CIs of Linear Model Coefficients for Species Occurrence and Detection Probalities
}
