% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{calc.pao}
\alias{calc.pao}
\title{Calculate Estimated Proportion of Sites At Which Species Occurs}
\usage{
calc.pao(topmod, stat = "mode", n)
}
\arguments{
\item{topmod}{Object of class \code{unmarkedFitOccu} for single-species occupancy model.}

\item{stat}{Statistic (\code{mean} or \code{median}) used to summarize posterior distribution.}

\item{n}{Number of sites surveyed.}
}
\value{
Vector containing posterior mean or median and upper and lower 95% confidence limits.
}
\description{
Calculate Estimated Proportion of Sites At Which Species Occurs
}
