% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{create.dir}
\alias{create.dir}
\title{Create a file directory for a new refuge I&M project}
\usage{
create.dir(proj.name = NULL, dir.name = NULL, moreFolders = NULL)
}
\arguments{
\item{proj.name}{The name of the project to be used as the name of the root project folder.}

\item{dir.name}{Location to create new root project folder. Default is the current working directory.}

\item{moreFolders}{List or vector containing character strings specifying full directory paths for additional
folders to be included in project directory.}
}
\description{
A function to create a new file directory structure. The directory structure is based off the Alaska I&M biometric project directory template.
}
