% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{create.raster}
\alias{create.raster}
\title{Create Spatial Raster Object Containing Simulated Habitat Covariate Values}
\usage{
create.raster(seed = 1234, ngr = 10)
}
\arguments{
\item{seed}{A single value used to set seed for replicability.}

\item{ngr}{A single value for number of rows and columns in site grid, e.g., ngr = 10 creates a 10x10 grid.}
}
\value{
A RasterLayer as from \code{\link{raster}}.
}
\description{
Create spatial raster object for a spatially correlated habitat covariate used to simulate species occurrence.
}
\examples{
create.raster(ngr = 10)
}
