% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{create.site.data}
\alias{create.site.data}
\title{Create year-specific site data for random sample of survey sites}
\usage{
create.site.data(year, habcov, nsites)
}
\arguments{
\item{year}{numeric value for survey year}

\item{habcov}{dataframe containing IDs, x-y coordinates, and habitat covariate values for all sites}

\item{nsites}{numeric value indicating the number of sites to randomly select from all sites that will be surveyed}
}
\value{
a dataframe containing IDs, x-y coordinates, and habitat covariate values for selected survey sites
}
\description{
Creates a site-level data object for sites to be surveyed conditional on year.
}
\details{
Note: year is used to set the seed for random site selection process.
}
\examples{
\dontrun{
rast <- create.rast()
habcov <- data.frame(id = as.factor(1:ncell(rast)), rasterToPoints(rast))
site.data <- create.site.data(year = 2020, habcov = habcov, nsites = 20)
str(site.data)
}

}
