% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{df2s123df}
\alias{df2s123df}
\title{Transform Occu.data Object to Survey123 Format}
\usage{
df2s123df(occu.data, year)
}
\arguments{
\item{occu.data}{Data.frame containing species occurrence and observation data.}

\item{year}{Numeric value for survey year.}
}
\value{
A data.frame formatted the same as data exported from Survey123.
}
\description{
Transforms a occu.data object to a data.frame that matches the format of data exported from Survey123.
}
