% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_functions.R
\name{extract.year}
\alias{extract.year}
\title{Extract Year From Data File Name}
\usage{
extract.year(filename)
}
\arguments{
\item{filename}{Full file path for data file.}
}
\value{
Character string.
}
\description{
Extract Year From Data File Name
}
