% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{get.betas}
\alias{get.betas}
\title{Extract Estimates of Model Beta Parameters}
\usage{
get.betas(mod)
}
\arguments{
\item{mod}{Object of class \code{unmarkedFitOccu} for single-species occupancy model.}
}
\value{
List of length 2 containing estimates of model parameters for
(1) occurrence probability and (2) detection probability
}
\description{
Extract Estimates of Model Beta Parameters
}
