% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{get.topmod}
\alias{get.topmod}
\title{Extract \code{unmarkedFitOccu} Object for Top Ranked Model}
\usage{
get.topmod(x1, x2)
}
\arguments{
\item{x1}{Year-specific \code{unmarkedFitList} object containing individual results for each model in a priori model set.}

\item{x2}{Year-specific vector containing model rankings.}
}
\value{
Object of class \code{unmarkedFitOccu}
}
\description{
Extract \code{unmarkedFitOccu} Object for Top Ranked Model
}
