% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_functions.R
\name{naive.occu}
\alias{naive.occu}
\title{Calculate Naive Species Occurrence for Surveyed Sites}
\usage{
naive.occu(x)
}
\arguments{
\item{x}{List of processed data objects created using the \code{process.data} function}
}
\value{
Vector containing values of 0 or 1 for each surveyed site indicating whether the species was observed (1) or unobserved (0).
}
\description{
Calculate Naive Species Occurrence for Surveyed Sites
}
