% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{one.step}
\alias{one.step}
\title{Simulate, Format, and Save Survey Site and Observation Data}
\usage{
one.step(
  year,
  habcov,
  nsites,
  nocc,
  p,
  showcoords,
  cex,
  ngood,
  npoor,
  write.data = FALSE,
  write.images = FALSE,
  return.data = FALSE
)
}
\arguments{
\item{year}{numeric value for survey year}

\item{habcov}{dataframe containing IDs, x-y coordinates, and habitat covariate values for all sites}

\item{nsites}{number of sites to randomly select from all sites that will be surveyed}

\item{nocc}{number of survey occasions conducted per site}

\item{p}{dataframe containing detection probabilities}

\item{showcoords}{logical value for whether to display coordinates for species location (TRUE)
or site name (FALSE) as main title}

\item{cex}{symbol size}

\item{ngood}{number of symbols used under good visibility conditions}

\item{npoor}{number of symbols used under poor visibility conditions}

\item{write.data}{logical value specifying whether to save site and occurrence data sets to raw data folder}

\item{write.images}{logical value specifying whether to save images for class exercise}
}
\value{
list of length 3 containing site.data, occu.data, and s123.data
}
\description{
Simulate, Format, and Save Survey Site and Observation Data
}
