% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{pao.plot}
\alias{pao.plot}
\title{Plot Annual Estimated Proportions of Sites At Which Species Occurs}
\usage{
pao.plot(PAO, years, modranks)
}
\arguments{
\item{PAO}{Data.frame containing posterior mean or median and upper and lower 95-percent confidence limits for all survey years.}

\item{years}{Numeric vector containing survey years.}

\item{modranks}{Data.frame containing model rankings for all survey years.}
}
\description{
Plot Annual Estimated Proportions of Sites At Which Species Occurs
}
