% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{predvals.plot}
\alias{predvals.plot}
\title{Plot Predicted Values for Probability of Species Occurrence and Detection}
\usage{
predvals.plot(predvals)
}
\arguments{
\item{predvals}{List of length 2 containing predicted values for probability of species
occurrence and probability of detection.}
}
\description{
Plot Predicted Values for Probability of Species Occurrence and Detection
}
