% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_functions.R
\name{process.data}
\alias{process.data}
\title{Process raw occurrence and site-level data for analysis}
\usage{
process.data(occudfs, sitedfs, years, habcov)
}
\arguments{
\item{occudfs}{list of dataframes containing annual raw species occurrence data}

\item{sitedfs}{list of dataframes containing annual raw site-level data}

\item{years}{numeric vector of survey years}

\item{habcov}{object of class \code{RasterLayer} containing spatial covariate}
}
\value{
list of length 6 with the following elements: 1) numeric vector of survey years, 2) list of year-specific
data objects of class unmarkedFrameOccu, 3) list of year-specific raw species occurrence data.frames, 4) list of
year-specific site-level data.frames, 5) list of year-specific numeric vectors containing naive occurrence values,
and 6) object of class RasterLayer containing a spatial covariate.
}
\description{
Process raw occurrence and site-level data for analysis
}
