% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report.R
\name{qc.report}
\alias{qc.report}
\title{Create a quality control report}
\usage{
qc.report(
  interface = T,
  o_dir = "C:/Users/mcobb/Desktop/r_webinar/docs/mccrea/demo/data/raw",
  s_dir = "C:/Users/mcobb/Desktop/r_webinar/docs/mccrea/demo/data/raw",
  hab_data = "./resources/data/geodata/habcov.asc",
  author = "Cliff Clavin",
  email = "cliff@fws.gov",
  phone = "555-867-5309",
  output_file = "report.pdf",
  output_dir = getwd(),
  ...
)
}
\arguments{
\item{interface}{if \code{TRUE}, user inputs the report parameters using a shiny interface}

\item{o_dir}{directory path to a folder containing the input occurrance Rdata files}

\item{s_dir}{directory path to a folder containing the input sample site Rdata files}

\item{hab_data}{an ASC raster file containing the input habitat covariate data}

\item{author}{the name of the report author}

\item{output_file}{output file name in \link{render}. Default is "report.html".}

\item{output_dir}{output directory for report in \link{render}. Default is user's current directory.}

\item{\dots}{other arguments to be passed to \link{render}.}
}
\value{

}
\description{
Create a quality control report
}
