% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{trial.plot}
\alias{trial.plot}
\title{Plot and Save Images for Class Exercise Using Survey123}
\usage{
trial.plot(
  good,
  z,
  path = NULL,
  fname = NULL,
  seed,
  color = NULL,
  ngood = 500,
  npoor = 1000,
  cex = 1,
  showcoords = FALSE,
  site = ""
)
}
\arguments{
\item{good}{logical value for whether visibility condition is good (TRUE) or poor (FALSE)}

\item{z}{indicator variable for whether site truly is occupied (1) or unoccupied (0)}

\item{path}{directory path for where image is saved}

\item{fname}{file name for image}

\item{seed}{value used to set seed in simulating x-y coordinates for the image}

\item{color}{color code used for symbol indicating species occurrence}

\item{ngood}{number of symbols used under good visibility conditions}

\item{npoor}{number of symbols used under poor visibility conditions}

\item{cex}{symbol size}

\item{showcoords}{logical value for whether to display coordinates for species location (TRUE)
or site name (FALSE) as main title}

\item{site}{character string specifiying site name}
}
\value{
No value is returned. Files are written to user-specified folder location.
}
\description{
Plot and Save Images for Class Exercise Using Survey123
}
