% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{annual.estimates}
\alias{annual.estimates}
\title{Create Plots of Annual Estimates of Species Occurrence and Detection Probability and LaTeX Code for Analysis Report}
\usage{
annual.estimates(Epsi, Ep, year, figdir.path)
}
\arguments{
\item{Epsi}{dataframe containing estimated values for occurrence probabilities, SE,
upper & lower CLs, and corresponding habitat covariate values}

\item{Ep}{dataframe containing estimated values for detection probabilities, SE,
upper & lower CLs, and corresponding visibility covariate values}

\item{year}{numeric value for survey year}

\item{figdir.path}{directory path for temporary storage of report figures}
}
\value{
List of length equal to number of survey years processed. Each element is contains a NULL value.
}
\description{
Estimates are based on top-ranked model and are conditional on estimated
relationships between model parameters and covariates when applicable.
}
