% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{annual.tables}
\alias{annual.tables}
\title{Create model selection and parameter estimate tables and LaTeX code for analysis report}
\usage{
annual.tables(modsel.table, year)
}
\arguments{
\item{modsel.table}{dataframe containing model output from \code{unmarked}}

\item{year}{numeric value for survey year}
}
\value{
List of length equal to number of survey years processed. Each element is contains a NULL value.
}
\description{
Create model selection and parameter estimate tables and LaTeX code for analysis report
}
