% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{create.raster}
\alias{create.raster}
\title{Create spatial raster object containing simulated habitat covariate values}
\usage{
create.raster(seed = 1234, ngr = 10)
}
\arguments{
\item{seed}{a single numeric value used to set seed for replicability}

\item{ngr}{a single numeric value indicating the number of rows and columns in site grid (e.g., ngr = 10 creates a 10x10 grid)}
}
\value{
a rasterLayer as from \code{\link{raster}}.
}
\description{
Creates spatial raster object for a spatially correlated habitat covariate used to simulate species occurrence
}
