% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{df2s123df}
\alias{df2s123df}
\title{Transform Occu.data Object to Survey123 Format}
\usage{
df2s123df(occu.data, year)
}
\arguments{
\item{occu.data}{dataframe containing species occurrence and observation data}

\item{year}{numeric value for survey year}
}
\value{
A dataframe formatted the same as data exported from ESRI Survey123
}
\description{
Transforms an \code{occu.data} object to a dataframe that matches the format of data exported from ESRI's Survey123
}
