% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_functions.R
\name{extract.year}
\alias{extract.year}
\title{Extract year From data file name}
\usage{
extract.year(filename)
}
\arguments{
\item{filename}{full file path to data file}
}
\value{
character string
}
\description{
Extract year From data file name
}
