% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_functions.R
\name{make.unmarkedData}
\alias{make.unmarkedData}
\title{Create formatted data object for analysis using \code{unmarked} package}
\usage{
make.unmarkedData(sitedf, occudf, year)
}
\arguments{
\item{sitedf}{dataframe containing raw site-level data}

\item{occudf}{dataframe containing raw species occurrence data}

\item{year}{numeric value for survey year}
}
\value{
Object of class \code{unamrkedFrameOccu}
}
\description{
Create formatted data object for analysis using \code{unmarked} package
}
