% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_functions.R
\name{make.unmarkedData}
\alias{make.unmarkedData}
\title{Create Formatted Data Object for Analysis Using unmarked Package}
\usage{
make.unmarkedData(sitedf, occudf, year)
}
\arguments{
\item{sitedf}{dataframe containing raw site-level data}

\item{occudf}{dataframe containing raw species occurrence data}

\item{year}{numeric value for survey year}
}
\value{
Object of class \code{unamrkedFrameOccu}
}
\description{
Create Formatted Data Object for Analysis Using unmarked Package
}
