% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{one.step}
\alias{one.step}
\title{Simulate, Format, and Save Survey Site and Observation Data}
\usage{
one.step(
  year,
  habcov,
  nsites,
  nocc,
  p,
  showcoords,
  cex,
  ngood,
  npoor,
  write.data = FALSE,
  write.images = FALSE,
  return.data = FALSE
)
}
\arguments{
\item{year}{Numeric value for survey year.}

\item{habcov}{Data.frame containing IDs, x-y coordinates, and habitat covariate values for all sites.}

\item{nsites}{Number of sites to randomly select from all sites that will be surveyed.}

\item{nocc}{Number of survey occasions conducted per site.}

\item{p}{Data.frame containing detection probabilities.}

\item{showcoords}{Logical value for whether to display coordinates for species location (TRUE)
or site name (FALSE) as main title.}

\item{cex}{Symbol size.}

\item{ngood}{Number of symbols used under good visibility conditions.}

\item{npoor}{Number of symbols used under poor visibility conditions.}

\item{write.data}{Logical value specifying whether to save site and occurrence data sets to raw data folder.}

\item{write.images}{Logical value specifying whether to save images for class exercise.}
}
\value{
List of length 3 containing site.data, occu.data, and s123.data.
}
\description{
Simulate, Format, and Save Survey Site and Observation Data
}
