% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{pao.plot}
\alias{pao.plot}
\title{Plot annual estimated proportions of sites at which a species occurs}
\usage{
pao.plot(PAO, years, modranks)
}
\arguments{
\item{PAO}{dataframe containing posterior mean or median and upper and lower 95-percent confidence limits for all survey years}

\item{years}{numeric vector containing survey years}

\item{modranks}{dataframe containing model rankings for all survey years}
}
\description{
Plot annual estimated proportions of sites at which a species occurs
}
