% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{predvals.plot}
\alias{predvals.plot}
\title{Plot predicted values for probability of species occurrence and detection}
\usage{
predvals.plot(predvals)
}
\arguments{
\item{predvals}{list of length 2 containing predicted values for probability of species
occurrence and probability of detection}
}
\description{
Plot predicted values for probability of species occurrence and detection
}
