% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_functions.R
\name{process.data}
\alias{process.data}
\title{Process Raw Occurrence and Site-Level Data for Analysis}
\usage{
process.data(occudfs, sitedfs, years, habcov)
}
\arguments{
\item{occudfs}{List of data.frames containing annual raw species occurrence data.}

\item{sitedfs}{List of data.frames containing annual raw site-level data.}

\item{years}{Numeric vector of survey years.}

\item{habcov}{Object of class RasterLayer containing spatial covariate.}
}
\value{
List of length 6 with the following elements: 1) numeric vector of survey years, 2) list of year-specific
data objects of class unmarkedFrameOccu, 3) list of year-specific raw species occurrence data.frames, 4) list of
year-specific site-level data.frames, 5) list of year-specific numeric vectors containing naive occurrence values,
and 6) object of class RasterLayer containing a spatial covariate.
}
\description{
Process Raw Occurrence and Site-Level Data for Analysis
}
