% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SppDistMonProj.R
\docType{package}
\name{SppDistMonProj}
\alias{SppDistMonProj}
\title{SppDistMonProj: a package for learning data workflow concepts with a Species Distribution Monitoring Project}
\description{
The \code{SppDistMonProj} package contains a suite of functions for suite of R
functions used for simulating, processing, and analyzing data from an
example survey project.
}
\section{Simulation functions}{

\itemize{
 \item{\code{create.raster}}
 \item{\code{create.site.data}}
 \item{\code{sim.occu.data}}
 \item{\code{df2s123df}}
 \item{\code{one.step}}
 }
}

\section{Quality control (QC) functions}{

The QC functions ...
}

\section{Analysis functions}{

The analysis functions ...
}

\section{Plot functions}{

The plot functions ...
}

\section{Misc functions}{

Other functions ...
}

