% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SppDistMonProj.R
\docType{package}
\name{SppDistMonProj}
\alias{SppDistMonProj}
\title{SppDistMonProj: a package for learning data workflow concepts with a Species Distribution Monitoring Project}
\description{
The SppDistMonProj package contains functions used simulate, quality check,
analyze, plot results and report for an hypothetical species distribution
monitoring project.
}
\section{Simulation functions}{

The simulation functions ...
}

\section{Quality control (QC) functions}{

The QC functions ...
}

\section{Analysis functions}{

The analysis functions ...
}

\section{Plot functions}{

The plot functions ...
}

\section{Misc functions}{

Other functions ...
}

